Added status message, restructuring
