+++ pkg/apis/migration/v1alpha1/gvk.go
@@ -6,3 +6,24 @@ type UnsupportedNamespace struct {
 	Name                 string   `json:"name"`
 	UnsupportedResources []string `json:"unsupportedResources"`
 }
+
+func (p *MigPlanStatus) SelectUnsupported() []string {
+	namespaces := []string{}
+	for _, ns := range p.UnsupportedNamespaces {
+		namespaces = append(namespaces, ns.Name)
+	}
+
+	return namespaces
+}
+
+func (p *MigPlanStatus) InitUnsupported() {
+	p.UnsupportedNamespaces = []UnsupportedNamespace{}
+}
+
+func (p *MigPlanStatus) AppendUnsupported(namespace string, resources []string) {
+	unsupportedNamespace := UnsupportedNamespace{
+		Name:                 namespace,
+		UnsupportedResources: resources,
+	}
+	p.UnsupportedNamespaces = append(p.UnsupportedNamespaces, unsupportedNamespace)
+}
