+++ pkg/controller/migplan/validation.go
@@ -54,6 +54,7 @@ const (
 	PvsDiscovered                  = "PvsDiscovered"
 	Closed                         = "Closed"
 	SourcePodsNotHealthy           = "SourcePodsNotHealthy"
+	NotSupported                   = "NotSupported"
 )
 
 // Categories
@@ -94,6 +95,7 @@ const (
 	StorageNotReadyMessage                = "The referenced `migStorageRef` does not have a `Ready` condition."
 	NsListEmptyMessage                    = "The `namespaces` list may not be empty."
 	InvalidDestinationClusterMessage      = "The `srcMigClusterRef` and `dstMigClusterRef` cannot be the same."
+	NsNotSupported                        = "Namespaces [] not supported. See: `unsupportedNamespaces` for details."
 	NsNotFoundOnSourceClusterMessage      = "Namespaces [] not found on the source cluster."
 	NsNotFoundOnDestinationClusterMessage = "Namespaces [] not found on the destination cluster."
 	NsLimitExceededMessage                = "Namespace limit: %d exceeded, found:%d."
